+++ frontend/public/CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Fixed
 
 - Update search bar to remove on-click animation for improved user experience
+- Fixed issue with plex not returning posters
 
 ---